requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.14
aioice==0.9.0
aiortc==1.10.1
aiosignal==1.3.2
attrs==25.3.0
av==13.1.0
bidict==0.23.1
blinker==1.9.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
contourpy==1.3.1
cryptography==44.0.2
cycler==0.12.1
dnspython==2.7.0
Flask==3.1.0
Flask-SocketIO==5.5.1
fonttools==4.56.0
frozenlist==1.5.0
google-crc32c==1.7.0
h11==0.14.0
idna==3.10
ifaddr==0.2.0
itsdangerous==2.2.0
Jinja2==3.1.6
kiwisolver==1.4.8
MarkupSafe==3.0.2
matplotlib==3.10.1
msgpack==1.1.0
multidict==6.2.0
numpy==2.2.4
packaging==24.2
pillow==11.1.0
propcache==0.3.0
pycparser==2.22
pyee==13.0.0
pylibsrtp==0.11.0
pyOpenSSL==25.0.0
pyparsing==3.2.3
pyrtlsdr[lib]
python-dateutil==2.9.0.post0
python-engineio==4.11.2
python-socketio==5.12.1
requests==2.32.3
scipy==1.15.2
setuptools==78.1.0
simple-websocket==1.1.0
six==1.17.0
typing_extensions==4.12.2
urllib3==2.3.0
websocket-client==1.8.0
Werkzeug==3.1.3
wsproto==1.2.0
yarl==1.18.3
flask>=2.0.0
python-dotenv>=0.19.0
gitpython>=3.1.0
